Stagger deleting a file based on filesize
When using sendfile in Crossbar, there can exist a race condition
where the filename has been passed to Cowboy for sending, Cowboy
handles streaming the file asynchronously, and calls Crossbar's
finish_request callback.

Because the streaming and finish_request callback are happening in
parallel, it is possible for the file to be deleted from the local
disk before streaming has completed.

pqc_cb_ledgers will occassionally provoke this race condition.

Since deleting the file occurs in a spawned process itself, use the
rounded natural log of the file's size * 10 to calculate a small sleep
timeout before deleting the file.